requirements.txt
bottle==0.12.8
peewee==2.5.0
PyJWT==1.0.0
redis==2.10.3
